Add files via upload
In the present scenario robots have become technological leaders. Although there is
rapid development in the field of agricultural robotic technologies, but it has been limited in the
fruit recognition and precision picking capabilities. In agriculture most of the manual work
involves at the harvesting stage. While harvesting humans commit lot of mistakes like plucking
over-ripen fruits and some unplucked fruits gets damaged by falling on ground which is a huge
loss. To overcome these problems we are designing a robot which will effectively work on
harvesting fruits. The main purpose of harvesting robot is to pluck the fruits from the trees
without damaging the fruit, and to reduce the burden of human and to speed up the process.
Color detection sensor and a robotic arm with gripper-cum-cutter are used to detect and pluck
fruits from the tree.